+++ HMCL/src/main/java/org/jackhuang/hmcl/ui/main/SettingsPage.java
@@ -30,7 +30,6 @@ import javafx.scene.control.ScrollPane;
 import javafx.scene.layout.BorderPane;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.VBox;
-import org.jackhuang.hmcl.Metadata;
 import org.jackhuang.hmcl.task.Schedulers;
 import org.jackhuang.hmcl.ui.Controllers;
 import org.jackhuang.hmcl.ui.FXUtils;
@@ -42,27 +41,13 @@ import org.jackhuang.hmcl.upgrade.UpdateChannel;
 import org.jackhuang.hmcl.upgrade.UpdateChecker;
 import org.jackhuang.hmcl.upgrade.UpdateHandler;
 import org.jackhuang.hmcl.util.Lang;
+import org.jackhuang.hmcl.util.LauncherLogExporter;
 import org.jackhuang.hmcl.util.StringUtils;
 import org.jackhuang.hmcl.util.i18n.I18n;
 import org.jackhuang.hmcl.util.i18n.SupportedLocale;
-import org.jackhuang.hmcl.util.io.FileUtils;
-import org.jackhuang.hmcl.util.io.IOUtils;
-import org.tukaani.xz.XZInputStream;
 
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 import java.util.concurrent.CompletableFuture;
-import java.util.zip.GZIPInputStream;
-import java.util.zip.ZipEntry;
-import java.util.zip.ZipOutputStream;
 
 import static org.jackhuang.hmcl.setting.SettingsManager.settings;
 import static org.jackhuang.hmcl.util.i18n.I18n.i18n;
@@ -221,7 +206,7 @@ public final class SettingsPage extends ScrollPane {
                     exportLogPane.setContent(logButton);
                     logButton.setOnAction(e -> {
                         exportLogPane.showSpinner();
-                        onExportLogs().whenCompleteAsync((result, exception) -> {
+                        CompletableFuture.supplyAsync(Lang.wrap(LauncherLogExporter::exportLogsAsZip), Schedulers.io()).whenCompleteAsync((result, exception) -> {
                             exportLogPane.hideSpinner();
                             if (exception == null) {
                                 Controllers.dialog(i18n("settings.launcher.launcher_log.export.success", result));
@@ -261,129 +246,5 @@ public final class SettingsPage extends ScrollPane {
             return;
         }
         UpdateHandler.updateFrom(target);
-    }
-
-    private static String getEntryName(Set<String> entryNames, String name) {
-        if (entryNames.add(name)) {
-            return name;
-        }
-
-        for (long i = 1; ; i++) {
-            String newName = name + "." + i;
-            if (entryNames.add(newName)) {
-                return newName;
-            }
-        }
-    }
-
-    /// This method guarantees to close both `input` and the current zip entry.
-    ///
-    /// If no exception occurs, this method returns `true`;
-    /// If an exception occurs while reading from `input`, this method returns `false`;
-    /// If an exception occurs while writing to `output`, this method will throw it as is.
-    private static boolean exportLogFile(ZipOutputStream output,
-                                         Path file, // For logging
-                                         String entryName,
-                                         InputStream input,
-                                         byte[] buffer) throws IOException {
-        //noinspection TryFinallyCanBeTryWithResources
-        try {
-            output.putNextEntry(new ZipEntry(entryName));
-            int read;
-            while (true) {
-                try {
-                    read = input.read(buffer);
-                    if (read <= 0)
-                        return true;
-                } catch (Throwable ex) {
-                    LOG.warning("Failed to decompress log file " + file, ex);
-                    return false;
-                }
-
-                output.write(buffer, 0, read);
-            }
-        } finally {
-            try {
-                input.close();
-            } catch (Throwable ex) {
-                LOG.warning("Failed to close log file " + file, ex);
-            }
-            output.closeEntry();
-        }
-    }
-
-    private CompletableFuture<Path> onExportLogs() {
-        return CompletableFuture.supplyAsync(Lang.wrap(() -> {
-            String nameBase = "hmcl-exported-logs-" + LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH-mm-ss"));
-            List<Path> recentLogFiles = LOG.findRecentLogFiles(5);
-
-            Path outputFile;
-            if (recentLogFiles.isEmpty()) {
-                outputFile = Metadata.CURRENT_DIRECTORY.resolve(nameBase + ".log");
-
-                LOG.info("Exporting latest logs to " + outputFile);
-                try (OutputStream output = Files.newOutputStream(outputFile)) {
-                    LOG.exportLogs(output);
-                }
-            } else {
-                outputFile = Metadata.CURRENT_DIRECTORY.resolve(nameBase + ".zip");
-
-                LOG.info("Exporting latest logs to " + outputFile);
-
-                byte[] buffer = new byte[IOUtils.DEFAULT_BUFFER_SIZE];
-                try (var os = Files.newOutputStream(outputFile);
-                     var zos = new ZipOutputStream(os)) {
-
-                    Set<String> entryNames = new HashSet<>();
-
-                    for (Path path : recentLogFiles) {
-                        String fileName = FileUtils.getName(path);
-                        String extension = StringUtils.substringAfterLast(fileName, '.');
-
-                        if ("gz".equals(extension) || "xz".equals(extension)) {
-                            // If an exception occurs while decompressing the input file, we should
-                            // ensure the input file and the current zip entry are closed,
-                            // then copy the compressed file content as-is into a new entry in the zip file.
-
-                            InputStream input = null;
-                            try {
-                                input = Files.newInputStream(path);
-                                input = "gz".equals(extension)
-                                        ? new GZIPInputStream(input)
-                                        : new XZInputStream(input);
-                            } catch (Throwable ex) {
-                                LOG.warning("Failed to open log file " + path, ex);
-                                IOUtils.closeQuietly(input, ex);
-                                input = null;
-                            }
-
-                            String entryName = getEntryName(entryNames, StringUtils.substringBeforeLast(fileName, "."));
-                            if (input != null && exportLogFile(zos, path, entryName, input, buffer))
-                                continue;
-                        }
-
-                        // Copy the log file content as-is into a new entry in the zip file.
-                        // If an exception occurs while decompressing the input file, we should
-                        // ensure the input file and the current zip entry are closed.
-
-                        InputStream input;
-                        try {
-                            input = Files.newInputStream(path);
-                        } catch (Throwable ex) {
-                            LOG.warning("Failed to open log file " + path, ex);
-                            continue;
-                        }
-
-                        exportLogFile(zos, path, getEntryName(entryNames, fileName), input, buffer);
-                    }
-
-                    zos.putNextEntry(new ZipEntry(getEntryName(entryNames, "hmcl-latest.log")));
-                    LOG.exportLogs(zos);
-                    zos.closeEntry();
-                }
-            }
-
-            return outputFile;
-        }), Schedulers.io());
     }
 }
